Support multivalued UWS parameters
Some UWS services (SODA, for example) have multivalued job parameters.
The parameter may be specified repeatedly with different values.
pydantic-xml supports this and produces a model element that holds
a list of Parameters, but it previously was not supported by vo-models
because the overridden `__init__` method didn't know how to handle
lists.

Add support to `Parameters.__init__` to first flatten all of the
input values into a simple list, regardless of which attribute
pydantic-xml thought they corresponded to, and then construct a dict
based on id values as before, but create a list of `Parameter` objects
if we see the same id more than once. Document that multi-valued
parameters should be declared as `MultiValuedParameter` instead of
`list[Parameter]`, which adds a validator that ensures that the
parameters are always converted to a list.

Add a test case for multivalued attributes and for correctly parsing
parameters given out of order in XML.

Fixes #21